+++ app/src/main/scala/vlthr/tee/expr.scala
@@ -136,10 +136,7 @@ final case class FilterExpr(expr: Expr,
                             filter: Filter,
                             args: List[Expr],
                             kwargs: Map[String, Expr])(
-    implicit val pctx: ParseContext,
-    hkArgs: HKernelAux[filter.Args],
-    ftArgs: FromTraversable[filter.Args],
-    ftOpt: FromTraversable[filter.OptArgs])
+    implicit val pctx: ParseContext)
     extends Expr {
   override def eval()(implicit ctx: Context) = {
     val exprVal = expr.eval()
